+++ packages/sampler/src/Environment/Environment.js
@@ -126,10 +126,7 @@ const StorybookDecorator = (story, config) => {
 		groupId: 'Development'
 	};
 
-	if (sample && sample.props && sample.props.info) {
-		config.description = sample.props.info;
-	}
-
+	const hasText = config.parameters && config.parameters.info && config.parameters.info.text;
 	const args = getArgs();
 	const classes = {
 		aria: boolean('debug aria', DevelopmentConfig, getKnobFromArgs(args, 'debug aria')),
@@ -143,8 +140,8 @@ const StorybookDecorator = (story, config) => {
 	return (
 		<PanelsBase
 			className={classnames(classes)}
-			title={`${config.kind} ${config.story}`.trim()}
-			description={config.description}
+			title={`${config.kind}`.replace(/\//g, ' ').trim()}
+			description={hasText ? config.parameters.info.text : null}
 			locale={select('locale', locales, Config)}
 			style={{
 				'--env-background': backgroundLabelMap[select('background', backgroundLabels, Config, getKnobFromArgs(args, 'background'))]