+++ packages/sampler/stories/default/VirtualGridList.js
@@ -3,7 +3,6 @@ import {boolean, number, select} from '@enact/storybook-utils/addons/knobs';
 import {ImageItem as UiImageItem} from '@enact/ui/ImageItem';
 import {mergeComponentMetadata} from '@enact/storybook-utils';
 import ri from '@enact/ui/resolution';
-import {storiesOf} from '@storybook/react';
 import {VirtualGridList, VirtualListBasic} from '@enact/ui/VirtualList';
 
 const
@@ -35,7 +34,7 @@ const
 
 const updateDataSize = (dataSize) => {
 	const
-		itemNumberDigits = dataSize > 0 ? ((dataSize - 1) + '').length : 0,
+		itemNumberDigits = dataSize > 0 ? (dataSize - 1 + '').length : 0,
 		headingZeros = Array(itemNumberDigits).join('0');
 
 	items.length = 0;
@@ -44,7 +43,7 @@ const updateDataSize = (dataSize) => {
 		const
 			count = (headingZeros + i).slice(-itemNumberDigits),
 			text = `Item ${count}${shouldAddLongContent({index: i, modIndex: 2})}`,
-			color = Math.floor((Math.random() * (0x1000000 - 0x101010)) + 0x101010).toString(16),
+			color = Math.floor(Math.random() * (0x1000000 - 0x101010) + 0x101010).toString(16),
 			source = `http://placehold.it/300x300/${color}/ffffff&text=Image ${i}`;
 
 		items.push({text, source});
@@ -57,31 +56,34 @@ updateDataSize(defaultDataSize);
 
 const VirtualGridListConfig = mergeComponentMetadata('VirtualGridList', VirtualListBasic, VirtualGridList);
 
-storiesOf('UI', module)
-	.add(
-		'VirtualList.VirtualGridList',
-		() => (
-			<VirtualGridList
-				dataSize={updateDataSize(number('dataSize', VirtualGridListConfig, defaultDataSize))}
-				direction={select('direction', prop.direction, VirtualGridListConfig)}
-				horizontalScrollbar={select('horizontalScrollbar', prop.scrollbarOption, VirtualGridListConfig)}
-				itemRenderer={uiRenderItem}
-				itemSize={{
-					minWidth: ri.scale(number('minWidth', VirtualGridListConfig, 180)),
-					minHeight: ri.scale(number('minHeight', VirtualGridListConfig, 270))
-				}}
-				key={select('scrollMode', prop.scrollModeOption, VirtualGridListConfig)}
-				noScrollByWheel={boolean('noScrollByWheel', VirtualGridListConfig)}
-				onScrollStart={action('onScrollStart')}
-				onScrollStop={action('onScrollStop')}
-				scrollMode={select('scrollMode', prop.scrollModeOption, VirtualGridListConfig)}
-				spacing={ri.scale(number('spacing', VirtualGridListConfig, 20))}
-				verticalScrollbar={select('verticalScrollbar', prop.scrollbarOption, VirtualGridListConfig)}
-			/>
-		),
-		{
-			info: {
-				text: 'Basic usage of VirtualGridList'
-			}
-		}
-	);
+export default {
+	title: 'UI/VirtualList.VirtualGridList',
+	component: 'VirtualGridList'
+};
+
+export const VirtualListVirtualGridList = () => (
+	<VirtualGridList
+		dataSize={updateDataSize(number('dataSize', VirtualGridListConfig, defaultDataSize))}
+		direction={select('direction', prop.direction, VirtualGridListConfig)}
+		horizontalScrollbar={select('horizontalScrollbar', prop.scrollbarOption, VirtualGridListConfig)}
+		itemRenderer={uiRenderItem}
+		itemSize={{
+			minWidth: ri.scale(number('minWidth', VirtualGridListConfig, 180)),
+			minHeight: ri.scale(number('minHeight', VirtualGridListConfig, 270))
+		}}
+		key={select('scrollMode', prop.scrollModeOption, VirtualGridListConfig)}
+		noScrollByWheel={boolean('noScrollByWheel', VirtualGridListConfig)}
+		onScrollStart={action('onScrollStart')}
+		onScrollStop={action('onScrollStop')}
+		scrollMode={select('scrollMode', prop.scrollModeOption, VirtualGridListConfig)}
+		spacing={ri.scale(number('spacing', VirtualGridListConfig, 20))}
+		verticalScrollbar={select('verticalScrollbar', prop.scrollbarOption, VirtualGridListConfig)}
+	/>
+);
+
+VirtualListVirtualGridList.storyName = 'VirtualList.VirtualGridList';
+VirtualListVirtualGridList.parameters = {
+	info: {
+		text: 'Basic usage of VirtualGridList'
+	}
+};