+++ packages/ui/ViewManager/ViewManager.js
@@ -70,10 +70,13 @@ const ViewManagerBase = class extends Component {
 		component: EnactPropTypes.renderable,
 
 		/**
-		 * Called with a reference to [component]{@link ui/ViewManager.ViewManager#component}
+		 * Called with a reference to the root component.
+		 *
+		 * When using {@link ui/ViewManager.ViewManager}, the `ref` prop is forwarded to this
+		 * component as `componentRef`.
 		 *
 		 * @type {Object|Function}
-		 * @private
+		 * @public
 		 */
 		componentRef: EnactPropTypes.ref,
 
@@ -316,6 +319,7 @@ const ViewManagerDecorator = ForwardRef({prop: 'componentRef'});
  * @memberof ui/ViewManager
  * @extends ui/ViewManager.ViewManagerBase
  * @mixes ui/ViewManager.ViewManagerDecorator
+ * @omit componentRef
  * @ui
  * @public
  */